requirements.txt
pyTelegramBotAPI
google-generativeai
python-dotenv
pyTelegramBotAPI
google-generativeai
weasyprint
gunicorn
requests
gevent